+++ test/conftest.py
@@ -1,13 +1,11 @@
 """Test HTML output the same way that Sphinx does in test_build_html.py."""
 
-import re
 from itertools import chain, cycle
 from pathlib import Path
 
 import pytest
 from docutils import nodes
 from lxml import etree as lxmltree
-from sphinx.testing.path import path as sphinx_path
 from sphinx.testing.util import SphinxTestApp
 
 pytest_plugins = 'sphinx.testing.fixtures'
@@ -17,7 +15,7 @@ etree_cache: dict[str, str] = {}
 
 @pytest.fixture(scope='session')
 def rootdir():
-    return sphinx_path(__file__).parent.abspath() / 'roots'
+    return Path(__file__).parent.absolute() / 'roots'
 
 
 class SphinxBuilder:
@@ -73,39 +71,3 @@ def cached_etree_parse():
 
 def flat_dict(d):
     return chain.from_iterable([zip(cycle([fname]), values) for fname, values in d.items()])
-
-
-def check_xpath(etree, fname, path, check, be_found=True):
-    nodes = list(etree.xpath(path))
-    if check is None:
-        assert nodes == [], f'found any nodes matching xpath {path!r} in file {fname}'
-        return
-    else:
-        assert nodes != [], f'did not find any node matching xpath {path!r} in file {fname}'
-    if callable(check):
-        check(nodes)
-    elif not check:
-        # only check for node presence
-        pass
-    else:
-
-        def get_text(node):
-            if node.text is not None:
-                # the node has only one text
-                return node.text
-            else:
-                # the node has tags and text; gather texts just under the node
-                return ''.join(n.tail or '' for n in node)
-
-        rex = re.compile(check)
-        if be_found:
-            if any(rex.search(get_text(node)) for node in nodes):
-                return
-        else:
-            if all(not rex.search(get_text(node)) for node in nodes):
-                return
-
-        raise AssertionError(
-            f'{check!r} not found in any node matching path {path} in {fname}: '
-            f'{[node.text for node in nodes]!r}'
-        )
